fix: post install script to enable systemd service on clean installs
The post installation script was missing the systemctl enable miru.service for clean installs while upgrades contained the enable command. This resulted in clean installs of miru not restarting on boot. The installation script has been refactored so that upgrades and clean installs use identical commands, both of which enable the sytemd service by default.